+++ MinArchDigsites.lua
@@ -100,7 +100,8 @@ function MinArch:UpdateActiveDigSites()
 				-- if not MinArchDigsitesGlobalDB["continent"][i][name]["race"] or MinArchDigsitesGlobalDB["continent"][i][name]["race"] == "Unknown" then
 					if MinArchDigsiteList[name] then
 						local race = GetArchaeologyRaceInfo(MinArchDigsiteList[name].race)
-						MinArchDigsitesGlobalDB["continent"][i][name]["race"] = race
+                        MinArchDigsitesGlobalDB["continent"][i][name]["race"] = race
+                        MinArchDigsitesGlobalDB["continent"][i][name]["raceId"] = MinArchDigsiteList[name].race
 					elseif not SpamBlock[name] then
 						ChatFrame1:AddMessage("Minimal Archaeology: Unknown digsite "..name)
 						SpamBlock[name] = 1
@@ -410,7 +411,7 @@ function MinArch:UpdateActiveDigSitesRace(Race)
 	end
 
 	-- print("UpdateActiveDigSitesRace", nearestDigSite, nearestDistance);
-	if (nearestDistance ~= nil and nearestDistance <= 5) then
+	if (nearestDistance ~= nil and nearestDistance <= MinArchDigsiteList[nearestDigSite].r * 1.1) then
 		MinArchDigsitesGlobalDB["continent"][tonumber(ContID)][nearestDigSite]["race"] = Race;
 		MinArchDigsitesGlobalDB["continent"][tonumber(ContID)][nearestDigSite]["zone"] = GetZoneText();
 		local subZone = GetSubZoneText();
@@ -427,7 +428,9 @@ end
 function MinArch:GetNearestDigsite()
 	if (IsInInstance()) then
 		return false;
-	end
+    end
+
+    local prioRace = MinArch.db.profile.TomTom.prioRace;
 
 	local nearestDistance = nil;
 	local nearestDigSite = nil;
@@ -448,23 +451,28 @@ function MinArch:GetNearestDigsite()
 
 	local continentID, worldPos = C_Map.GetWorldPosFromMapPos(uiMapID, playerPos);
 
-	ax = playerPos.x * 100;
-	ay = playerPos.y * 100;
+	ax = worldPos.x;
+	ay = worldPos.y;
 
 	for key, digsite in pairs(C_ResearchInfo.GetDigSitesForMap(uiMapID)) do
-		local name = tostring(digsite.name)
-		local digsitex = digsite.position.x * 100;
-		local digsitey = digsite.position.y * 100;
+        local name = tostring(digsite.name)
+        local _, digsiteWorldPos = C_Map.GetWorldPosFromMapPos(uiMapID, digsite.position)
+		local digsitex = digsiteWorldPos.x;
+		local digsitey = digsiteWorldPos.y;
 
 		local xd = math.abs(ax - tonumber(digsitex));
 		local yd = math.abs(ay - tonumber(digsitey));
 		local d = math.sqrt((xd*xd)+(yd*yd));
 
-		if (MinArchDigsitesDB["continent"][ContID][name] and MinArchDigsitesDB["continent"][ContID][name]["status"] == true) then
-			if (nearestDigSite == nil or d < nearestDistance) then
-				nearestDigSite = name;
-				nearestDistance = d;
-				nearestDigSiteDetails = MinArchDigsitesGlobalDB["continent"][ContID][nearestDigSite];
+        if (MinArchDigsitesDB["continent"][ContID][name] and MinArchDigsitesDB["continent"][ContID][name]["status"] == true) then
+            local currentRace = MinArchDigsiteList[name].race;
+            if ( (prioRace == currentRace and (nearestDistance == nil or nearestDigSiteDetails.raceId ~= prioRace or (nearestDigSiteDetails.raceId == prioRace and d < nearestDistance) ) )
+                or nearestDigSite == nil
+                or (nearestDigSiteDetails.raceId ~= prioRace and d < nearestDistance))
+            then
+                nearestDigSite = name;
+                nearestDistance = d;
+                nearestDigSiteDetails = MinArchDigsitesGlobalDB["continent"][ContID][nearestDigSite];
 			end
 		end
 	end