+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 - New feataure: **Companion** (**distance tracker** and more, details down below)
 - Fixed issues caused by Shadowlands LUA changes
 - MinArch is now compatible with addons that emulate TomTom (like Carbonite Maps)
+- Added an option to prioritize a selected race when creating automatic waypoints (Selectable in the TomTom section of MinArch settings)
+- Right clicking on the auto-waypoint button now opens the waypoint settings page
+- Fixed an issue with the automatic waypoints sometimes not selecting the closest digsite
 
 ### 8.3.0